+++ lib/views/calibration_form.dart
@@ -52,10 +52,7 @@ class _CalibrationFormState extends State<CalibrationForm> {
             children: [
               SizedBox(
                   width: 110,
-                  child: Text(
-                      isAzimuth
-                          ? S.of(context).azimuthRange
-                          : S.of(context).elevationRange,
+                  child: Text(isAzimuth ? S.of(context).azimuthRange : S.of(context).elevationRange,
                       style: Theme.of(context).textTheme.titleSmall)),
               const SizedBox(width: 10),
               SizedBox(
@@ -65,9 +62,7 @@ class _CalibrationFormState extends State<CalibrationForm> {
                     decoration: InputDecoration(labelText: S.of(context).begin),
                     textAlign: TextAlign.right,
                     controller: _rangeBeginController,
-                    inputFormatters: [
-                      FilteringTextInputFormatter.allow(RegExp('[0-9]'))
-                    ],
+                    inputFormatters: [FilteringTextInputFormatter.allow(RegExp('[0-9]'))],
                     validator: _rangeValidator,
                   )),
               const SizedBox(height: 40, child: Text("°")),
@@ -79,19 +74,14 @@ class _CalibrationFormState extends State<CalibrationForm> {
                     decoration: InputDecoration(labelText: S.of(context).end),
                     textAlign: TextAlign.right,
                     controller: _rangeEndController,
-                    inputFormatters: [
-                      FilteringTextInputFormatter.allow(RegExp('[0-9]'))
-                    ],
+                    inputFormatters: [FilteringTextInputFormatter.allow(RegExp('[0-9]'))],
                     validator: _rangeValidator,
                   )),
               const SizedBox(height: 40, child: Text("°")),
               const SizedBox(width: 10),
               SizedBox(
                   width: 120,
-                  child: Text(
-                      isAzimuth
-                          ? S.of(context).horizontalSpeed
-                          : S.of(context).verticalSpeed,
+                  child: Text(isAzimuth ? S.of(context).horizontalSpeed : S.of(context).verticalSpeed,
                       style: Theme.of(context).textTheme.titleSmall)),
               const SizedBox(width: 10),
               SizedBox(
@@ -101,9 +91,7 @@ class _CalibrationFormState extends State<CalibrationForm> {
                       textAlign: TextAlign.center,
                       decoration: const InputDecoration(labelText: 'ms/°'),
                       controller: _speedController,
-                      inputFormatters: [
-                        FilteringTextInputFormatter.allow(RegExp('[0-9]'))
-                      ],
+                      inputFormatters: [FilteringTextInputFormatter.allow(RegExp('[0-9]'))],
                       validator: (value) {
                         if (value == null || value.trim().isEmpty) {
                           return S.of(context).fillIn;
@@ -123,9 +111,7 @@ class _CalibrationFormState extends State<CalibrationForm> {
                   width: 200,
                   child: ElevatedButton(
                       onPressed: rotatorModel.isConnected ? _onReset : null,
-                      child: Text(isAzimuth
-                          ? S.of(context).horizontalReset
-                          : S.of(context).verticalReset))),
+                      child: Text(isAzimuth ? S.of(context).horizontalReset : S.of(context).verticalReset))),
               const SizedBox(width: 10),
               SizedBox(
                   width: 200,
@@ -133,10 +119,7 @@ class _CalibrationFormState extends State<CalibrationForm> {
                       onPressed: rotatorModel.isConnected ? _onCalibrate : null,
                       child: Text(S.of(context).startCalibrate))),
               const SizedBox(width: 10),
-              SizedBox(
-                  width: 200,
-                  child: ElevatedButton(
-                      onPressed: _onSave, child: Text(S.of(context).save))),
+              SizedBox(width: 200, child: ElevatedButton(onPressed: _onSave, child: Text(S.of(context).save))),
             ],
           )
         ],
@@ -157,8 +140,7 @@ class _CalibrationFormState extends State<CalibrationForm> {
 
   void _calibrateFinish() {
     RotatorModel rotatorModel = context.read();
-    int duration =
-        DateTime.now().difference(rotatorModel.moveStartTime).inMilliseconds;
+    int duration = DateTime.now().difference(rotatorModel.moveStartTime).inMilliseconds;
 
     if (isAzimuth) {
       rotatorModel.horizontalSpeed = duration ~/ rotatorModel.azimuthRange;
@@ -183,9 +165,7 @@ class _CalibrationFormState extends State<CalibrationForm> {
   void _onResetFinish() {
     RotatePanTiltCommand().run(RotatorAction.stop);
     RotatorModel model = context.read();
-    isAzimuth
-        ? model.currentAzimuth = model.azimuthBegin.toDouble()
-        : model.currentElevation = model.elevationBegin.toDouble();
+    isAzimuth ? model.currentAzimuth = model.azimuthBegin : model.currentElevation = model.elevationBegin;
     setState(() => _resetInProgress = false);
   }
 
@@ -228,8 +208,7 @@ class _CalibrationFormState extends State<CalibrationForm> {
     rotatorModel.horizontalSpeed = speed;
   }
 
-  void _saveElevation(
-      RotatorModel rotatorModel, int begin, int end, int speed) {
+  void _saveElevation(RotatorModel rotatorModel, int begin, int end, int speed) {
     rotatorModel.elevationBegin = begin;
     rotatorModel.elevationEnd = end;
     rotatorModel.verticalSpeed = speed;
